# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: SuperCore API
  version: 1.0.0
  description: |
    # SuperCore - Universal Entity Management Platform API

    SuperCore is a revolutionary meta-platform for creating Core Banking systems through natural language.

    ## Core Concepts

    - **Oracle**: Platform consciousness that knows its identity, licenses, and capabilities
    - **Object Definitions**: Blueprints (DNA) for all entities in the system
    - **Instances**: Living entities created from object definitions
    - **Relationships**: Connections between instances forming a semantic graph
    - **Validation Rules**: Reusable business rules and compliance checks

    ## Authentication

    Currently in development. Future versions will use Logto for OAuth 2.0/OIDC.

    ## Rate Limiting

    Default: 100 requests per minute per IP

    ## Versioning

    API version is included in the URL path: `/api/v1/`

  contact:
    name: SuperCore Team
    email: dev@lbpay.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api-dev.supercore.lbpay.com/api/v1
    description: Development environment
  - url: https://api.supercore.lbpay.com/api/v1
    description: Production environment
...
paths:
...
  /object-definitions:
    get:
      tags: [Object Definitions]
      summary: List all object definitions
      description: Returns a paginated list of object definitions
      operationId: listObjectDefinitions
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [BUSINESS_ENTITY, RULE, POLICY, INTEGRATION, LOGIC]
          description: Filter by category
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of object definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ObjectDefinition'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Object Definitions]
      summary: Create a new object definition
      description: Creates a new object definition (blueprint)
      operationId: createObjectDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectDefinition'
            examples:
              conta_corrente:
                summary: Conta Corrente (Checking Account)
                value:
                  name: conta_corrente
                  display_name: Conta Corrente PF
                  description: Conta corrente para pessoa física
                  category: BUSINESS_ENTITY
                  schema:
                    type: object
                    required: [cpf, nome, saldo]
                    properties:
                      cpf:
                        type: string
                        pattern: "^[0-9]{11}$"
                      nome:
                        type: string
                        minLength: 3
                      saldo:
                        type: number
                        minimum: 0
                  states:
                    initial: DRAFT
                    states: [DRAFT, ACTIVE, SUSPENDED, CLOSED]
                    transitions:
                      - from: DRAFT
                        to: ACTIVE
                      - from: ACTIVE
                        to: SUSPENDED
                      - from: SUSPENDED
                        to: ACTIVE
                      - from: ACTIVE
                        to: CLOSED
      responses:
        '201':
          description: Object definition created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /object-definitions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Object Definitions]
      summary: Get object definition by ID
      operationId: getObjectDefinition
      responses:
        '200':
          description: Object definition found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '404':
          description: Object definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Object Definitions]
      summary: Update object definition
      operationId: updateObjectDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateObjectDefinition'
      responses:
        '200':
          description: Object definition updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '404':
          description: Object definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Object Definitions]
      summary: Delete object definition
      description: Soft delete an object definition (marks as inactive)
      operationId: deleteObjectDefinition
      responses:
        '204':
          description: Object definition deleted
        '404':
          description: Object definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /object-definitions/{id}/schema:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Object Definitions]
      summary: Get JSON Schema for object definition
      description: Returns only the JSON Schema portion for form generation
      operationId: getObjectDefinitionSchema
      responses:
        '200':
          description: JSON Schema
          content:
            application/json:
              schema:
                type: object
                description: JSON Schema Draft 7

  # ============================================================================
  # INSTANCES ENDPOINTS
  # ============================================================================
...
components:
  schemas:
...
    ObjectDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: conta_corrente
        display_name:
          type: string
          example: Conta Corrente PF
        description:
          type: string
        version:
          type: integer
        schema:
          type: object
          description: JSON Schema Draft 7
        rules:
          type: array
          items:
            type: object
        states:
          type: object
        ui_hints:
          type: object
        relationships:
          type: array
        category:
          type: string
          enum: [BUSINESS_ENTITY, RULE, POLICY, INTEGRATION, LOGIC]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean
    CreateObjectDefinition:
      type: object
      required: [name, display_name, schema, category]
      properties:
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        schema:
          type: object
        rules:
          type: array
        states:
          type: object
        ui_hints:
          type: object
        relationships:
          type: array
        category:
          type: string
          enum: [BUSINESS_ENTITY, RULE, POLICY, INTEGRATION, LOGIC]
    UpdateObjectDefinition:
      type: object
      properties:
        display_name:
          type: string
        description:
          type: string
        schema:
          type: object
        rules:
          type: array
        states:
          type: object
        ui_hints:
          type: object
        relationships:
          type: array
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object